requirements.txt
requests
moviepy
pydub
google-api-python-client
google-auth-oauthlib
google-auth
Pillow
python-dotenv
replicate
transformers
opencv-python
tqdm
openai
gTTS==2.5.4
imageio
imageio-ffmpeg
